resolve https://issues.apache.org/activemq/browse/AMQ-2889 - fix setter type
git-svn-id: https://svn.apache.org/repos/asf/activemq/trunk@991102 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-ra/src/main/java/org/apache/activemq/ra/ActiveMQConnectionRequestInfo.java
@@ -213,9 +213,9 @@ public class ActiveMQConnectionRequestInfo implements ConnectionRequestInfo, Ser
         return Boolean.valueOf(redeliveryPolicy().isUseExponentialBackOff());
     }
 
-    public void setRedeliveryBackOffMultiplier(Short value) {
+    public void setRedeliveryBackOffMultiplier(Double value) {
         if (value != null) {
-            redeliveryPolicy().setBackOffMultiplier(value.shortValue());
+            redeliveryPolicy().setBackOffMultiplier(value);
         }
     }
